s3_service/requirements.txt
fastapi[standard]
pydantic
boto3
python-dotenv